Refactoring: Split WeatherSensor.cpp into 3 files
